LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-zh-Hant-HK.html

<!DOCTYPE html><!-- webkit-test-runner [ language=zh-Hant-HK ] -->
<html>
<head>
<script src="../../../../../resources/js-test.js"></script>
</head>
<body>
<script>
shouldBeEqualToString("navigator.language", "zh-hk");
</script>
</body>
</html>
